requirements.txt
asgiref==3.8.1
Django==5.1.1
django-rest==0.8.7
djangorestframework==3.15.2
pillow==10.4.0
pip==23.0.1
setuptools==65.5.0
psycopg2==2.9.3
six==1.16.0
sqlparse==0.5.1
python-decouple==3.4
typing_extensions==4.12.2